combine warrant date validation tests into single cypress test
